backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pandas==2.2.2
python-dotenv==1.0.1
